Some interface additions, use eightfold reducer (for cellular automaton mode) as library
I made the interface changes a long while ago, but they were for replaying after games ended in the GUI mode.
The reducer.index() function (with the Pólya enumeration) isn't currently used, because the dict seems faster, but setting dictMode=False may be the only way (though I'm thinking I will make a specialised program for this purpose using it (that computes periodicities on-the-fly instead of regressing the state transition diagram)).